bloodfy/requirements.txt
# Project Requirements
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
python-dotenv>=1.0.0

# Database
mysqlclient>=2.2.0

# Task Queue
celery>=5.3.0
redis>=5.0.0

# External APIs
twilio>=8.10.0
googlemaps>=4.10.0
google-cloud-dialogflow>=2.28.0

# Geolocation & AI
geopy>=2.4.0
numpy>=1.24.0
pandas>=2.0.0

# Utilities
Pillow>=10.0.0
python-dateutil>=2.8.0
